Refactor to core pubsub
